refactor api calls
