--- src/services/mumble.ts
@@ -1,58 +1,0 @@
-import { decodeTime } from 'ulid';
-import { User } from './users';
-import axios from 'axios';
-
-export type Mumble = {
-  id: string;
-  creator: string;
-  creatorProfile?: User;
-  text: string;
-  mediaUrl: string;
-  mediaType: string;
-  likeCount: number;
-  likedByUser: boolean;
-  type: string;
-  replyCount: number;
-  createdTimestamp: number;
-  createdDate?: string;
-};
-
-type RawMumble = Omit<Mumble, 'createdTimestamp'>;
-
-type QwackerMumbleResponse = {
-  count: number;
-  data: RawMumble[];
-};
-
-export const fetchMumbles = async (params?: { limit?: number; offset?: number; newerThanMumbleId?: string }) => {
-  const { limit, offset, newerThanMumbleId } = params || {};
-
-  const url = `${process.env.NEXT_PUBLIC_QWACKER_API_URL}/posts?${new URLSearchParams({
-    limit: limit?.toString() || '10',
-    offset: offset?.toString() || '0',
-    newerThan: newerThanMumbleId || '',
-  })}`;
-
-  const res = await fetch(url, {
-    headers: {
-      'content-type': 'application/json',
-    },
-  });
-  const { count, data } = (await res.json()) as QwackerMumbleResponse;
-
-  const mumbles = data.map(transformMumble);
-
-  return {
-    count,
-    mumbles,
-  };
-};
-
-// @ts-ignore
-const transformMumble = (mumble: RawMumble) => ({
-  ...mumble,
-  createdTimestamp: decodeTime(mumble.id),
-  createdDate: new Date(decodeTime(mumble.id)).toLocaleDateString(),
-});
-
-
