+++ src/services/posts.ts
@@ -1,5 +1,26 @@
-import { transformMumble, UploadImage } from './serviceTypes';
+import { QwackerMumbleResponse, transformMumble, UploadImage } from './serviceTypes';
 import axios from 'axios';
+
+export const fetchMumbles = async (params?: { limit?: number; offset?: number; newerThanMumbleId?: string }) => {
+  const { limit, offset, newerThanMumbleId } = params || {};
+
+  const url = `${process.env.NEXT_PUBLIC_QWACKER_API_URL}/posts?${new URLSearchParams({
+    limit: limit?.toString() || '10',
+    offset: offset?.toString() || '0',
+    newerThan: newerThanMumbleId || '',
+  })}`;
+
+  const res = await axios.get(url);
+
+  const { count, data } = res.data as QwackerMumbleResponse;
+
+  const mumbles = data.map(transformMumble);
+
+  return {
+    count,
+    mumbles,
+  };
+};
 
 export const postMumble = async (text: string, file: UploadImage | null, accessToken?: string) => {
   if (!accessToken) {
