+++ src/services/serviceTypes.ts
@@ -1,4 +1,5 @@
 import { decodeTime } from 'ulid';
+import { User } from './users';
 
 export type UploadImage = File & { preview: string };
 
@@ -9,3 +10,25 @@ export const transformMumble = (mumble: RawMumble) => ({
   createdTimestamp: decodeTime(mumble.id),
   createdDate: new Date(decodeTime(mumble.id)).toLocaleDateString(),
 });
+
+export type Mumble = {
+  id: string;
+  creator: string;
+  creatorProfile?: User;
+  text: string;
+  mediaUrl: string;
+  mediaType: string;
+  likeCount: number;
+  likedByUser: boolean;
+  type: string;
+  replyCount: number;
+  createdTimestamp: number;
+  createdDate?: string;
+};
+
+export type RawMumble = Omit<Mumble, 'createdTimestamp'>;
+
+export type QwackerMumbleResponse = {
+  count: number;
+  data: RawMumble[];
+};
